! First stab at displaying read/unread state for a department, so I'll explain it. It is slightly fuzzy but hopefully it'll make sense.
We start out by getting the last message posted in a department, and the last message we're known to have read in a given department - if the first is higher than the second, there's something in that department we haven't read. If, however, we don't touch the helpdesk for too long for whatever reason, it falls off after a week which is primarily a performance limiter. So if you don't look at the helpdesk for a week and nothing changes in that time it shows up as read anyway.

Try it out, see how well it works in reality/practicality, it should function mostly analoguous to SMF boards.